:recycle: UUID apply and Image package create
