+++ src/main/kotlin/com/highv/ecommerce/common/image/dto/ImageUrlResponse.kt
@@ -1,4 +1,4 @@
-package com.highv.ecommerce.domain.auth.dto
+package com.highv.ecommerce.common.image.dto
 
 data class ImageUrlResponse(
     val imageUrl: String
