+++ src/main/kotlin/com/highv/ecommerce/domain/auth/controller/UserController.kt
@@ -5,7 +5,6 @@ import com.highv.ecommerce.common.dto.DefaultResponse
 import com.highv.ecommerce.common.exception.ValidationException
 import com.highv.ecommerce.domain.auth.dto.EmailAuthRequest
 import com.highv.ecommerce.domain.auth.dto.EmailAuthResponse
-import com.highv.ecommerce.domain.auth.dto.ImageUrlResponse
 import com.highv.ecommerce.domain.auth.dto.LoginRequest
 import com.highv.ecommerce.domain.auth.service.UserService
 import com.highv.ecommerce.infra.security.UserPrincipal
@@ -67,17 +66,4 @@ class UserController(
             .body(userService.verifyCode(request.email, request.role, code))
     }
 
-    @PreAuthorize("hasRole('ADMIN') or hasRole('SELLER') or hasRole('BUYER')")
-    @PostMapping("/image")
-    fun uploadImage(
-        @RequestPart file: MultipartFile,
-        @AuthenticationPrincipal user: UserPrincipal
-    ): ResponseEntity<ImageUrlResponse> = ResponseEntity.ok(userService.uploadImage(file, user.id))
-
-    @PreAuthorize("hasRole('ADMIN') or hasRole('SELLER') or hasRole('BUYER')")
-    @PostMapping("/images")
-    fun uploadImages(
-        @RequestPart files: List<MultipartFile>,
-        @AuthenticationPrincipal user: UserPrincipal
-    ): ResponseEntity<List<ImageUrlResponse>> = ResponseEntity.ok(userService.uploadImages(files, user.id))
 }
