+++ src/main/kotlin/com/highv/ecommerce/domain/auth/service/UserService.kt
@@ -7,7 +7,6 @@ import com.highv.ecommerce.common.exception.CustomRuntimeException
 import com.highv.ecommerce.common.exception.EmailAlreadyExistsException
 import com.highv.ecommerce.common.exception.SellerLoginFailedException
 import com.highv.ecommerce.domain.auth.dto.EmailAuthResponse
-import com.highv.ecommerce.domain.auth.dto.ImageUrlResponse
 import com.highv.ecommerce.domain.auth.dto.LoginRequest
 import com.highv.ecommerce.domain.auth.dto.UserRole
 import com.highv.ecommerce.domain.buyer.entity.Buyer
@@ -154,31 +153,6 @@ class UserService(
         }
 
         return false
-    }
-
-    fun uploadImage(file: MultipartFile, id: Long): ImageUrlResponse {
-        s3Manager.uploadFile(file)
-        val imageUrl = s3Manager.getFile(file.originalFilename)
-        return ImageUrlResponse(imageUrl = imageUrl)
-    }
-
-    fun uploadImages(files: List<MultipartFile>, id: Long): List<ImageUrlResponse> {
-
-        if (files.size > 9) {
-            throw CustomRuntimeException(409, "이미지는 최대 9장만 등록 가능합니다.")
-        } else if (files.isEmpty()) {
-            throw CustomRuntimeException(409, "이미지를 등록해주세요")
-        }
-
-        val imageUrls: MutableList<ImageUrlResponse> = mutableListOf()
-
-        files.forEach {
-            s3Manager.uploadFile(it)
-            val imageUrl = s3Manager.getFile(it.originalFilename)
-            imageUrls.add(ImageUrlResponse(imageUrl = imageUrl))
-        }
-
-        return imageUrls
     }
 
     companion object {
